added action bar to today view, many other minor tweaks
